#9274- refactorings
